Split out client code into its own module.
